sns: Swappable SNS client getter
Makes it possible to provide an alternative SNS client getter in tests.
