+++ pkg/reconciler/awssnssource/subscription_test.go
@@ -21,168 +21,13 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 
-	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/client-go/kubernetes/fake"
 
 	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/service/sns"
-
-	"github.com/triggermesh/aws-event-sources/pkg/apis/sources/v1alpha1"
 )
-
-func TestAwsCredentials(t *testing.T) {
-	const (
-		ns = "fake-namespace"
-
-		accessKeyIDKey     = "key-id"
-		accessKeyIDVal     = "fake key ID"
-		secretAccessKeyKey = "secret-key"
-		secretAccessKeyVal = "fake secret"
-	)
-
-	testCases := []struct {
-		name        string
-		initSecrets []*corev1.Secret
-		input       v1alpha1.AWSSecurityCredentials
-		expect      *credentials.Value
-		getRequests int
-	}{
-		{
-			name: "Both from value",
-			input: v1alpha1.AWSSecurityCredentials{
-				AccessKeyID: v1alpha1.ValueFromField{
-					Value: accessKeyIDVal,
-				},
-				SecretAccessKey: v1alpha1.ValueFromField{
-					Value: secretAccessKeyVal,
-				},
-			},
-			expect: &credentials.Value{
-				AccessKeyID:     accessKeyIDVal,
-				SecretAccessKey: secretAccessKeyVal,
-			},
-			getRequests: 0,
-		},
-		{
-			name: "One from value, the other from secret",
-			initSecrets: []*corev1.Secret{
-				newSecret(ns, "secret1", map[string]string{
-					secretAccessKeyKey: secretAccessKeyVal,
-				}),
-			},
-			input: v1alpha1.AWSSecurityCredentials{
-				AccessKeyID: v1alpha1.ValueFromField{
-					Value: accessKeyIDVal,
-				},
-				SecretAccessKey: v1alpha1.ValueFromField{
-					ValueFromSecret: &corev1.SecretKeySelector{
-						LocalObjectReference: corev1.LocalObjectReference{
-							Name: "secret1",
-						},
-						Key: secretAccessKeyKey,
-					},
-				},
-			},
-			expect: &credentials.Value{
-				AccessKeyID:     accessKeyIDVal,
-				SecretAccessKey: secretAccessKeyVal,
-			},
-			getRequests: 1,
-		},
-		{
-			name: "Both from same secret",
-			initSecrets: []*corev1.Secret{
-				newSecret(ns, "secret1", map[string]string{
-					accessKeyIDKey:     accessKeyIDVal,
-					secretAccessKeyKey: secretAccessKeyVal,
-				}),
-			},
-			input: v1alpha1.AWSSecurityCredentials{
-				AccessKeyID: v1alpha1.ValueFromField{
-					ValueFromSecret: &corev1.SecretKeySelector{
-						LocalObjectReference: corev1.LocalObjectReference{
-							Name: "secret1",
-						},
-						Key: accessKeyIDKey,
-					},
-				},
-				SecretAccessKey: v1alpha1.ValueFromField{
-					ValueFromSecret: &corev1.SecretKeySelector{
-						LocalObjectReference: corev1.LocalObjectReference{
-							Name: "secret1",
-						},
-						Key: secretAccessKeyKey,
-					},
-				},
-			},
-			expect: &credentials.Value{
-				AccessKeyID:     accessKeyIDVal,
-				SecretAccessKey: secretAccessKeyVal,
-			},
-			getRequests: 1,
-		},
-		{
-			name: "Both from a different secret",
-			initSecrets: []*corev1.Secret{
-				newSecret(ns, "secret1", map[string]string{
-					accessKeyIDKey: accessKeyIDVal,
-				}),
-				newSecret(ns, "secret2", map[string]string{
-					secretAccessKeyKey: secretAccessKeyVal,
-				}),
-			},
-			input: v1alpha1.AWSSecurityCredentials{
-				AccessKeyID: v1alpha1.ValueFromField{
-					ValueFromSecret: &corev1.SecretKeySelector{
-						LocalObjectReference: corev1.LocalObjectReference{
-							Name: "secret1",
-						},
-						Key: accessKeyIDKey,
-					},
-				},
-				SecretAccessKey: v1alpha1.ValueFromField{
-					ValueFromSecret: &corev1.SecretKeySelector{
-						LocalObjectReference: corev1.LocalObjectReference{
-							Name: "secret2",
-						},
-						Key: secretAccessKeyKey,
-					},
-				},
-			},
-			expect: &credentials.Value{
-				AccessKeyID:     accessKeyIDVal,
-				SecretAccessKey: secretAccessKeyVal,
-			},
-			getRequests: 2,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			secrets := make([]runtime.Object, len(tc.initSecrets))
-			for i, secret := range tc.initSecrets {
-				secrets[i] = secret
-			}
-
-			cli := fake.NewSimpleClientset(secrets...)
-
-			creds, err := awsCredentials(cli.CoreV1().Secrets(ns), &tc.input)
-
-			require.NoError(t, err)
-
-			assert.Equal(t, tc.expect.AccessKeyID, creds.AccessKeyID, "Value of access key ID")
-			assert.Equal(t, tc.expect.SecretAccessKey, creds.SecretAccessKey, "Value of secret access key")
-			assert.Equal(t, tc.getRequests, len(cli.Actions()), "Number of API requests")
-		})
-	}
-}
 
 func TestErrors(t *testing.T) {
 	genericErr := assert.AnError
@@ -216,20 +61,4 @@ func TestErrors(t *testing.T) {
 		assert.False(t, isNotFound(genericK8SErr))
 		assert.False(t, isNotFound(genericErr))
 	})
-}
-
-func newSecret(ns, name string, data map[string]string) *corev1.Secret {
-	secr := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: ns,
-			Name:      name,
-		},
-		Data: make(map[string][]byte, len(data)),
-	}
-
-	for k, v := range data {
-		secr.Data[k] = []byte(v)
-	}
-
-	return secr
 }
